remove test file that does not related to main file
